requirements.txt
ultralytics==8.0.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
facenet-pytorch==2.5.3
